refactor: Adds InewsFtpTimestampParser interface.
