+++ src/business-logic/interfaces/inews-ftp-timestamp-parser.ts
@@ -1,0 +1,3 @@
+export interface InewsFtpTimestampParser {
+  parseInewsFtpTimestamp(timestampText: string): number
+}
